refactor: reorganize UI component imports and update paths
- Updated TypeScript configuration to change the path for UI components from 'src/components/ui' to 'src/shared/ui'.
- Refactored imports across multiple files to reflect the new path structure.
- Removed obsolete UI component files to streamline the codebase.
