+++ src/app/layout.config.tsx
@@ -1,8 +1,8 @@
 import { siteConfig } from '@/core/site-config'
 import type { BaseLayoutProps } from 'fumadocs-ui/layouts/shared'
 import { Logo } from '@/components/logo'
-import { CodeThemeSwitch } from '@/components/ui/code-theme-switch'
-import { ThemeSwitcher } from '@/components/ui/theme-switcher'
+import { CodeThemeSwitch } from '@/shared/ui/code-theme-switch'
+import { ThemeSwitcher } from '@/shared/ui/theme-switcher'
 
 /**
  * Shared layout configurations
