--- src/components/ui/index.ts
@@ -1,6 +1,0 @@
-export * from '../ui/separator'
-export * from '../ui/button'
-export * from '../ui/card'
-export * from '../ui/input'
-export * from '../ui/tabs'
-export * from '../ui/tooltip'
