+++ src/shared/ui/code-theme-switch.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useTheme } from 'next-themes';
-import { Button } from '../../shared/ui/button';
+import { Button } from './button';
 import { Moon, Sun } from 'lucide-react';
 import { useEffect, useState } from 'react';
