+++ src/shared/ui/index.ts
@@ -1,0 +1,6 @@
+export * from './separator'
+export * from './button'
+export * from './card'
+export * from './input'
+export * from './tabs'
+export * from './tooltip'
